PL-9900 AmCharts startDuration should take double value

+++ modules/gui/src/com/haulmont/charts/gui/amcharts/model/HasStartEffect.java
@@ -10,6 +10,6 @@ public interface HasStartEffect<T> {
     AnimationEffect getStartEffect();
     T setStartEffect(AnimationEffect startEffect);
 
-    Integer getStartDuration();
-    T setStartDuration(Integer startDuration);
+    Double getStartDuration();
+    T setStartDuration(Double startDuration);
 }
