+++ modules/web/src/com/haulmont/charts/web/gui/components/charts/amcharts/WebCoordinateChart.java
@@ -68,12 +68,12 @@ public abstract class WebCoordinateChart<T extends CoordinateChart,
     }
 
     @Override
-    public Integer getStartDuration() {
+    public Double getStartDuration() {
         return getModel().getStartDuration();
     }
 
     @Override
-    public T setStartDuration(Integer startDuration) {
+    public T setStartDuration(Double startDuration) {
         getModel().setStartDuration(startDuration);
         return (T) this;
     }
